refactor: [GSW-2025] refactor separate hooks by responsibility
