+++ packages/web/src/layouts/pool/pool-remove/containers/remove-liquidity-container/RemoveLiquidityContainer.tsx
@@ -1,7 +1,7 @@
 import React, { useCallback, useMemo, useState } from "react";
 
 import useCustomRouter from "@hooks/common/use-custom-router";
-import { usePositionData } from "@hooks/common/use-position-data";
+import { usePositionData } from "@hooks/pool/data/use-position-data";
 import { useWallet } from "@hooks/wallet/data/use-wallet";
 
 import RemoveLiquidity from "../../components/remove-liquidity/RemoveLiquidity";
