+++ packages/web/src/layouts/wallet/components/wallet-my-positions-header/WalletMyPositionsHeader.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { useTranslation } from "react-i18next";
 
-import { usePositionData } from "@hooks/common/use-position-data";
+import { usePositionData } from "@hooks/pool/data/use-position-data";
 import { useWallet } from "@hooks/wallet/data/use-wallet";
 
 import { wrapper } from "./WalletMyPositionsHeader.styles";
