+++ packages/web/src/layouts/wallet/components/wallet-my-positions/WalletMyPositions.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-import { usePositionData } from "@hooks/common/use-position-data";
+import { usePositionData } from "@hooks/pool/data/use-position-data";
 import { useWallet } from "@hooks/wallet/data/use-wallet";
 
 import { wrapper } from "./WalletMyPositions.styles";
