+++ packages/web/src/layouts/wallet/containers/asset-list-container/AssetListContainer.tsx
@@ -7,7 +7,7 @@ import AssetReceiveModal from "@components/wallet/asset-receive-modal/AssetRecei
 import useClickOutside from "@hooks/common/use-click-outside";
 import useCustomRouter from "@hooks/common/use-custom-router";
 import { useLoading } from "@hooks/common/use-loading";
-import { usePositionData } from "@hooks/common/use-position-data";
+import { usePositionData } from "@hooks/pool/data/use-position-data";
 import { usePreventScroll } from "@hooks/common/use-prevent-scroll";
 import { useWindowSize } from "@hooks/common/use-window-size";
 import { useTokenData } from "@hooks/token/data/use-token-data";
