+++ packages/web/src/layouts/wallet/containers/wallet-balance-container/WalletBalanceContainer.tsx
@@ -7,8 +7,8 @@ import AssetReceiveModal from "@components/wallet/asset-receive-modal/AssetRecei
 import { WRAPPED_GNOT_PATH } from "@constants/environment.constant";
 import { useBroadcastHandler } from "@hooks/common/use-broadcast-handler";
 import { useMessage } from "@hooks/common/use-message";
-import { usePosition } from "@hooks/common/use-position";
-import { usePositionData } from "@hooks/common/use-position-data";
+import { usePosition } from "@hooks/pool/data/use-position";
+import { usePositionData } from "@hooks/pool/data/use-position-data";
 import { usePreventScroll } from "@hooks/common/use-prevent-scroll";
 import { useTransactionConfirmModal } from "@hooks/common/use-transaction-confirm-modal";
 import { useTransactionEventStore } from "@hooks/common/use-transaction-event-store";
